whisper-server/requirements.txt
flask==3.0.3
flask-cors==4.0.0
openai-whisper==20231117
torch==2.2.2
numpy==1.26.4